pyproject.toml
[build-system]
requires = ["setuptools>=67", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sof-discord-bot"
version = "0.1.0"
description = "Discord webhook scoreboard bot for Soldier of Fortune (SoF1) servers"
readme = "README.md"
authors = [{name = "SoF Community", email = "noreply@example.com"}]
requires-python = ">=3.9"
dependencies = [
  "Pillow>=10.0.0",
  "requests>=2.31.0",
  "watchdog>=4.0.0",
  "python-dotenv>=1.0.1"
]
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
]

[project.urls]
Homepage = "https://github.com/your-org/sof-discord-bot"

[project.scripts]
sof-discord-bot = "sof_discord_bot.cli:main"
